+++ kinto_signer/updater.py
@@ -201,8 +201,14 @@ class LocalUpdater(object):
         return self._get_records(self.destination)
 
     def push_records_to_destination(self, request):
-        __, timestamp = self.get_destination_records()
-        new_records, _ = self.get_source_records(last_modified=timestamp)
+        __, dest_timestamp = self.get_destination_records()
+        new_records, source_timestamp = self.get_source_records(last_modified=dest_timestamp)
+
+        if source_timestamp and dest_timestamp and dest_timestamp > source_timestamp:
+            raise ValueError("Destination collection timestamp cannot be higher "
+                             "than source collection timestamp. Check that your "
+                             "storage backend timezone is UTC.")
+
         # Update the destination collection.
         for record in new_records:
             storage_kwargs = {