+++ tests/test_updater.py
@@ -95,9 +95,17 @@ class LocalUpdaterTest(unittest.TestCase):
                    return_value=([], 1324))
         records = [{'id': idx, 'foo': 'bar %s' % idx} for idx in range(1, 4)]
         self.patch(self.updater, 'get_source_records',
-                   return_value=(records, '42'))
+                   return_value=(records, 1325))
         self.updater.push_records_to_destination(DummyRequest())
         assert self.storage.update.call_count == 3
+
+    def test_push_records_to_destination_raises_if_storage_is_misconfigured(self):
+        self.patch(self.updater, 'get_destination_records',
+                   return_value=([], 1324))
+        self.patch(self.updater, 'get_source_records',
+                   return_value=([], 1234))
+        with pytest.raises(ValueError):
+            self.updater.push_records_to_destination(DummyRequest())
 
     def test_push_records_removes_deleted_records(self):
         self.patch(self.updater, 'get_destination_records',
@@ -106,7 +114,7 @@ class LocalUpdaterTest(unittest.TestCase):
         records.extend([{'id': idx, 'deleted': True, 'last_modified': 42}
                         for idx in range(3, 5)])
         self.patch(self.updater, 'get_source_records',
-                   return_value=(records, '42'))
+                   return_value=(records, 1325))
         self.updater.push_records_to_destination(DummyRequest())
         self.updater.get_source_records.assert_called_with(last_modified=1324)
         assert self.storage.update.call_count == 2
@@ -122,7 +130,7 @@ class LocalUpdaterTest(unittest.TestCase):
         records.extend([{'id': idx, 'deleted': True, 'last_modified': 42}
                        for idx in range(3, 5)])
         self.patch(self.updater, 'get_source_records',
-                   return_value=(records, '42'))
+                   return_value=(records, 1325))
         # Calling the updater should not raise the RecordNotFoundError.
         self.updater.push_records_to_destination(DummyRequest())
 
@@ -131,7 +139,7 @@ class LocalUpdaterTest(unittest.TestCase):
                    return_value=([], None))
         records = [{'id': idx, 'foo': 'bar %s' % idx} for idx in range(1, 4)]
         self.patch(self.updater, 'get_source_records',
-                   return_value=(records, '42'))
+                   return_value=(records, 1325))
         self.updater.push_records_to_destination(DummyRequest())
         self.updater.get_source_records.assert_called_with(last_modified=None)
         assert self.storage.update.call_count == 3